+++ Engine/Geometry/Curves/Incomplete/Cosine.cs
@@ -9,7 +9,6 @@
 
 using Engine.Imaging;
 using System;
-using System.Drawing;
 
 namespace Engine.Geometry
 {
@@ -94,7 +93,7 @@ namespace Engine.Geometry
         /// <remarks></remarks>
         public static Point2D Interpolate(Point2D a, Point2D b, double index)
         {
-            //Single MU2 = (float)((1.0 - Math.Cos(index * 180)) * 0.5);
+            //Single MU2 = (double)((1.0 - Math.Cos(index * 180)) * 0.5);
             //return Y1 * (1.0 - MU2) + Y2 * MU2;
             double MU2 = (1.0 - Math.Cos(index * 180)) * 0.5;
             return (Point2D)a.Scale(1.0 - MU2).Add(b.Scale(MU2));
@@ -111,8 +110,8 @@ namespace Engine.Geometry
         {
             double MU = ((1 - Math.Cos((Index * 180))) / 2);
             return new Point2D(
-                (float)((a.X * (1 - MU)) + (b.X * MU)),
-                (float)((a.Y * (1 - MU)) + (b.Y * MU))
+                (a.X * (1 - MU)) + (b.X * MU),
+                (a.Y * (1 - MU)) + (b.Y * MU)
                 );
         }
